Inform the user of the most specific tag for a SHA
Previously, the error message would only show the SHA of the
un-sha-pinned action. Now, the most specific tag is shown as well,
which will allow the user to quickly add a semver comment to the
workflow.

"Most specific", here, means the longest (in terms of characters) tag
that points to the discovered SHA.